Refactor imports to use analytics module instead of dataAnalysis across test and component files
